move: current classes moved to new folder
