+++ lib/cli/version.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 module RailsInteractive
-  VERSION = "1.0.0"
+  class CLI
+    VERSION = "1.0.0"
+  end
 end
